+++ src/common/xml-traverser.ts
@@ -1,9 +1,9 @@
 import { Position } from "vscode";
 import * as slimdom from "slimdom";
-import * as slimdom2 from "slimdom-sax-parser";
+
 export class XmlTraverser {
 
-    constructor(private _xmlDocument: slimdom2.Document) { }
+    constructor(private _xmlDocument: slimdom.Document) { }
 
     get xmlDocument(): slimdom.Document {
         return this._xmlDocument;
@@ -32,11 +32,7 @@ export class XmlTraverser {
 
     }
 
-    getElementAtPosition(position: Position): slimdom.Element {
-        const node = this.getNodeAtPosition(position);
-
-        return this.getNearestElementAncestor(node);
-    }
+   
 
     getNearestElementAncestor(node: slimdom.Node): slimdom.Element {
         if (!this.isElement) {
@@ -46,86 +42,15 @@ export class XmlTraverser {
         return <slimdom.Element>node;
     }
 
-    getNodeAtPosition(position: Position): slimdom.Node {
-        return this._getNodeAtPositionCore(position, this._xmlDocument.documentElement);
-    }
+ 
 
-    getSiblings(node: slimdom.Node): slimdom.Node[] {
-        if (this.isElement(node)) {
-            return this.getSiblingElements(node);
-        }
-
-        return this.getSiblingAttributes(node);
-    }
-
-    getSiblingAttributes(node: slimdom.Node): slimdom.Node[] {
-        return this.getChildAttributeArray(<slimdom.Element>node.parentNode);
-    }
-
-    getSiblingElements(node: slimdom.Node): slimdom.Node[] {
-        return this.getChildElementArray(node.parentNode);
-    }
-
-    hasSimilarSiblings(node: slimdom.Node): boolean {
-        if (!node || !node.parentNode || !this.isElement(node)) {
-            return false;
-        }
-
-        const siblings = this.getChildElementArray(<slimdom.Element>node.parentNode);
-
-        return (siblings.filter(x => x.tagName === (node as slimdom.Element).tagName).length > 1);
-    }
+  
 
     isElement(node: slimdom.Node): boolean {
         return (!!node && !!(node as slimdom.Element).tagName);
     }
 
-    private _getNodeAtPositionCore(position: Position, contextNode: slimdom.Node): slimdom.Node {
-        if (!contextNode) {
-            return undefined;
-        }
-
-        const lineNumber = contextNode.lineNumber;
-        const columnNumber = (contextNode as any).columnNumber;
-        const columnRange = [columnNumber, (columnNumber + (this._getNodeWidthInCharacters(contextNode) - 1))];
-
-        // for some reason, xmldom sets the column number for attributes to the "="
-        if (!this.isElement(contextNode)) {
-            columnRange[0] = (columnRange[0] - contextNode.nodeName.length);
-        }
-
-        if (this._checkRange(lineNumber, position, columnRange)) {
-            return contextNode;
-        }
-
-        if (this.isElement(contextNode)) {
-            // if the element contains text, check to see if the cursor is present in the text
-            const textContent = (contextNode as slimdom.Element).textContent;
-
-            if (textContent) {
-                columnRange[1] = (columnRange[1] + textContent.length);
-
-                if (this._checkRange(lineNumber, position, columnRange)) {
-                    return contextNode;
-                }
-            }
-
-            const children = [...this.getChildAttributeArray(<slimdom.Element>contextNode), ...this.getChildElementArray(contextNode)];
-            let result: slimdom.Node;
-
-            for (let i = 0; i < children.length; i++) {
-                const child = children[i];
-
-                result = this._getNodeAtPositionCore(position, child);
-
-                if (result) {
-                    return result;
-                }
-            }
-        }
-
-        return undefined;
-    }
+ 
 
     private _checkRange(lineNumber: number, position: Position, columnRange: number[]): boolean {
         return (lineNumber === (position.line + 1) && ((position.character + 1) >= columnRange[0] && (position.character + 1) < columnRange[1]));